+++ lib/ModelMixins/GeojsonMixin.ts
@@ -1,5 +1,5 @@
 import i18next from "i18next";
-import { computed, observable, runInAction } from "mobx";
+import { action, computed, observable, runInAction } from "mobx";
 import Cartesian3 from "terriajs-cesium/Source/Core/Cartesian3";
 import Color from "terriajs-cesium/Source/Core/Color";
 import defaultValue from "terriajs-cesium/Source/Core/defaultValue";
@@ -29,6 +29,13 @@ import UrlMixin from "../ModelMixins/UrlMixin";
 import Model from "../Models/Model";
 import proxyCatalogItemUrl from "../Models/proxyCatalogItemUrl";
 import { GeoJsonTraits } from "../Traits/GeoJsonTraits";
+import { Feature, FeatureCollection, GeoJsonObject } from "geojson";
+import DiscretelyTimeVaryingMixin, {
+  DiscreteTimeAsJS
+} from "./DiscretelyTimeVaryingMixin";
+import TimeIntervalCollection from "terriajs-cesium/Source/Core/TimeIntervalCollection";
+import TimeInterval from "terriajs-cesium/Source/Core/TimeInterval";
+import Iso8601 from "terriajs-cesium/Source/Core/Iso8601";
 
 const formatPropertyValue = require("../Core/formatPropertyValue");
 const hashFromString = require("../Core/hashFromString");
@@ -52,7 +59,9 @@ type Coordinates = number[];
 export default function GeoJsonMixin<
   T extends Constructor<Model<GeoJsonTraits>>
 >(Base: T) {
-  abstract class GeoJsonMixin extends MappableMixin(UrlMixin(Base)) {
+  abstract class GeoJsonMixin extends DiscretelyTimeVaryingMixin(
+    MappableMixin(UrlMixin(Base))
+  ) {
     protected readonly zipFileRegex = /(\.zip\b)/i;
 
     readonly canZoomTo = true;
@@ -150,6 +159,52 @@ export default function GeoJsonMixin<
         });
     }
 
+    @action
+    private addPerPropertyStyleToGeoJson(json: JsonObject | GeoJsonObject) {
+      const geojson = json as GeoJsonObject;
+      if (geojson.type === "Feature") {
+        const featureProperties = (geojson as Feature).properties;
+        if (featureProperties === null) {
+          return;
+        }
+        const featurePropertiesEntires = Object.entries(featureProperties);
+
+        const matchedStyles = this.perPropertyStyles.filter(style => {
+          const stylePropertiesEntries = Object.entries(
+            style.properties as any
+          );
+
+          // For every key-value pair in the style, is there an identical one in the feature's properties?
+          return stylePropertiesEntries.every(
+            ([styleKey, styleValue]) =>
+              featurePropertiesEntires.find(([featKey, featValue]) => {
+                if (typeof styleValue === "string" && !style.caseSensitive) {
+                  featKey === styleKey &&
+                    (featValue as string).toLowerCase() ===
+                      (styleValue as string).toLowerCase();
+                }
+                return featKey === styleKey && featValue === styleValue;
+              }) !== undefined
+          );
+        });
+
+        if (matchedStyles !== undefined) {
+          for (let matched of matchedStyles) {
+            for (let trait of Object.keys(matched.style.traits)) {
+              featureProperties[trait] =
+                // @ts-ignore - TS can't tell that `trait` is of the correct index type for style
+                matched.style[trait] ?? featureProperties[trait];
+            }
+          }
+        }
+      } else if (geojson.type === "FeatureCollection") {
+        const featureCollection = geojson as FeatureCollection;
+        featureCollection.features.forEach(feature => {
+          this.addPerPropertyStyleToGeoJson(feature);
+        });
+      }
+    }
+
     private loadDataSource(geoJson: JsonObject): Promise<GeoJsonDataSource> {
       /* Style information is applied as follows, in decreasing priority:
              - simple-style properties set directly on individual features in the GeoJSON file
@@ -158,6 +213,8 @@ export default function GeoJsonMixin<
              - if anything is underspecified there, then Cesium's defaults come in.
              See https://github.com/mapbox/simplestyle-spec/tree/master/1.1.0
           */
+
+      this.addPerPropertyStyleToGeoJson(geoJson);
 
       function defaultColor(
         colorString: string | undefined,
@@ -232,12 +289,42 @@ export default function GeoJsonMixin<
 
       return makeRealPromise<GeoJsonDataSource>(
         GeoJsonDataSource.load(geoJson, options)
-      ).then(function(dataSource) {
+      ).then(dataSource => {
         const entities = dataSource.entities;
         for (let i = 0; i < entities.values.length; ++i) {
           const entity = entities.values[i];
 
           const properties = entity.properties;
+
+          // Time
+          if (
+            isDefined(properties) &&
+            isDefined(this.timeProperty) &&
+            isDefined(this.discreteTimesAsSortedJulianDates)
+          ) {
+            const startTimeDiscreteTime = properties[this.timeProperty];
+            const startTimeIdx = this.discreteTimesAsSortedJulianDates?.findIndex(
+              t => t.tag === startTimeDiscreteTime.getValue()
+            );
+            const startTime = this.discreteTimesAsSortedJulianDates[
+              startTimeIdx
+            ];
+
+            if (isDefined(startTime)) {
+              const endTimeIdx = startTimeIdx + 1;
+              const endTime = this.discreteTimesAsSortedJulianDates[endTimeIdx];
+
+              entity.availability = new TimeIntervalCollection([
+                new TimeInterval({
+                  start: startTime.time,
+                  stop: endTime?.time ?? Iso8601.MAXIMUM_VALUE,
+                  isStopIncluded: false
+                })
+              ]);
+            }
+          }
+
+          // Billboard
           if (isDefined(entity.billboard) && isDefined(options.markerUrl)) {
             entity.billboard = new BillboardGraphics({
               image: new ConstantProperty(options.markerUrl),
@@ -353,6 +440,40 @@ export default function GeoJsonMixin<
         }
         return dataSource;
       });
+    }
+
+    @computed
+    get discreteTimes(): DiscreteTimeAsJS[] | undefined {
+      if (this.timeProperty === undefined || this.readyData === undefined) {
+        return undefined;
+      }
+      const discreteTimes: DiscreteTimeAsJS[] = [];
+      const addFeatureToDiscreteTimes = (geojson: GeoJsonObject) => {
+        if (geojson.type === "Feature") {
+          let feature = geojson as Feature;
+          if (
+            feature.properties !== null &&
+            feature.properties !== undefined &&
+            feature.properties[this.timeProperty!] !== undefined
+          ) {
+            discreteTimes.push({
+              time: new Date(
+                `${feature.properties[this.timeProperty!]}`
+              ).toISOString(),
+              tag: feature.properties[this.timeProperty!]
+            });
+          }
+        } else if (geojson.type === "FeatureCollection") {
+          const featureCollection = geojson as FeatureCollection;
+          featureCollection.features.forEach(feature =>
+            addFeatureToDiscreteTimes(feature)
+          );
+        }
+      };
+
+      addFeatureToDiscreteTimes((this.readyData as unknown) as GeoJsonObject);
+
+      return discreteTimes;
     }
 
     protected abstract async customDataLoader(